Added files via upload
The changes on version 2 include slight tweaks in visuals and minor changes visually, however a lot of mistakes in the HTML sheet was found especially since i was using center elements and align in html when i should be using them in the CSS style sheet. Several other issues were detected with the help of a HTML validator program and have been dealt with.